refactor(store-core): split types.ts into domain files behind a barrel
Break the 1008-line types.ts into cohesive domain files (chat / session /
model / connection / settings / server-signals / git / conversation)
re-exported through src/types/index.ts.

Pure move: the public surface is byte-identical (all 48 exports preserved),
and every `from './types'` / `from '../types'` import now resolves to the
directory index unchanged. App + dashboard consume only the @chroxy/store-core
barrel, so no consumer import changes. Enables tree-shaking and keeps each
domain independently readable.
